backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
emails==0.6
pydantic-settings==2.1.0
email-validator==2.1.0
pymongo==4.6.1
motor==3.3.2
aiofiles
cloudinary==1.37.0
fastapi-mail==1.4.1
PyJWT==2.8.0